chore - extract suggest status widget
